Move game state logging to gameGlobalInfo.
