+++ src/epsilonServer.cpp
@@ -1,7 +1,6 @@
 #include "epsilonServer.h"
 #include "playerInfo.h"
 #include "gameGlobalInfo.h"
-#include "gameStateLogger.h"
 #include "main.h"
 
 EpsilonServer::EpsilonServer()
@@ -14,15 +13,10 @@ EpsilonServer::EpsilonServer()
     engine->setGameSpeed(0.0);
     for(unsigned int n=0; n<factionInfo.size(); n++)
         factionInfo[n]->reset();
-
-    state_logger = new GameStateLogger();
-    state_logger->start();
 }
 
 EpsilonServer::~EpsilonServer()
 {
-    if (state_logger)
-        state_logger->destroy();
 }
 
 void EpsilonServer::onNewClient(int32_t client_id)
